move pc to page_compiler (#22)
