+++ internal/page_compiler/emitter.go
@@ -1,0 +1,7 @@
+package page_compiler
+
+type PageEmitter struct{}
+
+func NewPageEmitter() *PageEmitter {
+	return &PageEmitter{}
+}
