--- internal/pc/page.go
@@ -1,1 +1,0 @@
-package pc
